+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/dynamictext/card/Item.kt
@@ -3,7 +3,7 @@ package cn.spacexc.wearbili.dataclass.dynamic.dynamictext.card
 data class Item(
     val content: String,
     val ctrl: String,
-    val orig_dy_id: Int,
+    val orig_dy_id: Long,
     val pre_dy_id: Int,
     val reply: Int,
     val rp_id: Long,
